requirements.txt
absl-py==2.2.2
ale-py==0.10.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.1
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
gymnasium==1.1.1
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
#nvidia-cublas-cu12==12.4.5.8
#nvidia-cuda-cupti-cu12==12.4.127
#nvidia-cuda-nvrtc-cu12==12.4.127
#nvidia-cuda-runtime-cu12==12.4.127
#nvidia-cudnn-cu12==9.1.0.70
#nvidia-cufft-cu12==11.2.1.3
#nvidia-curand-cu12==10.3.5.147
#nvidia-cusolver-cu12==11.6.1.9
#nvidia-cusparse-cu12==12.3.1.170
#nvidia-cusparselt-cu12==0.6.2
#nvidia-nccl-cu12==2.21.5
#nvidia-nvjitlink-cu12==12.4.127
#nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
protobuf==5.29.4
psutil==7.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
sentry-sdk==2.26.1
setproctitle==1.3.5
shtab==1.7.2
six==1.17.0
smmap==5.0.2
stable_baselines3==2.6.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typeguard==4.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tyro==0.9.18
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.9
Werkzeug==3.1.3
